Bugfixing

+++ src/main/java/JBKRMobile/OwnedStock.java
@@ -2,9 +2,9 @@ package JBKRMobile;
 
 public class OwnedStock {
     private String ticker;
-    private long quantity;
+    private int quantity;
 
-    public OwnedStock(String ticker, long quantity) {
+    public OwnedStock(String ticker, int quantity) {
         this.ticker = ticker;
         this.quantity = quantity;
     }
@@ -14,20 +14,22 @@ public class OwnedStock {
         return ticker;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
     public void addQuantity(int amount) {
         quantity += amount;
-    }  
+    }
 
     /**
-     * Subtracts an int from quantity. If amount is greater than quantity, then returns false before subtracting. 
+     * Subtracts an int from quantity. If amount is greater than quantity, then
+     * returns false before subtracting.
      * Otherwise, returns true
+     * 
      * @param amount int
      */
-    public boolean subtractQuantity (int amount) {
+    public boolean subtractQuantity(int amount) {
         if (quantity - amount < 0) {
             return false;
         }
